fix deviceTest failures
Some tests are failing due to an NPE in BaseTask.emptyFolder method.
This is due to the fact the plugin instance variable is null.
The plugin instance is null since it is not set by the users of the AndroidReportTask
and in particular the ReportingPlugin which is not subclassing BasePlugin.

Once fix would be to make the ReportingPlugin subclass BasePlugin and make sure
the plugin instance variable is set.

Quick fix to not display a logging message when the plugin instance is not set.

Change-Id: I894c72f4eac94fafcab42ff85d89d9867a229b26