80115: Set correct compressed sizes in resource shrinking APK
The JarEntry(JarEntry) constructor copies all the JarEntry
entries, including compressed size, even though the actual
compressed size may differ. This causes an exception when the
entry is closed.

This CL fixes this such that it only sets the JarEntry size
fields for STORED (not DEFLATED) files.

It also changes the compressed level to 9 (from
Deflater.DEFAULT_COMPRESSION) since that seems to match the
better compression levels used in existing .ap_ files.)

I was able to reproduce the crash with the iosched application.
However, when I also improve the compression level to 9, the
discrepancy goes away, so I don't have a new resource file
in the unit test to reproduce the crash. However, the build
test for the resource shrinker continues to assert that the
remaining entries in the compressed .ap_ matches the ones
in the unstripped .ap_.

Change-Id: Icdf79861d98dfa894f0dc45f3ac3fbbee93ec30d