Don't accumulate outputs while pre-dexing, keep the last result only.
When building multiple subprojects in parallel, we might end up racing to
pre-dex the same jar. While this is benign, we should only persist a single
result in the dex cache -- and we used to do this until mOutputFiles became
a list. If we store duplicates, we break the next incremental build while
trying to copy the same file over.

Bug: https://code.google.com/p/android/issues/detail?id=75580
Change-Id: Idced6a80066e0bebe9bcabd61a8d6c56ec6a33a5