Handle local jar in test for multi-dex libraries.
In several places we have to manually exclude local jar
dependencies of libraries in their tests. This is because we
don't properly remove these dependencies from the direct test
dependencies, and they also show up in the aar bundles.

This is one more place where we need to do this. This should be
cleaned up with the upcoming work on dependencies.

Change-Id: Icab18da3e384516d926841f9b31fd9cd39944cbc