80872: Resource shrinking does not shrink as much as it could
For some reason, recent versions of the Gradle plugin (or ProGuard)
ends up with all the R$type classes from libraries in the final ProGuard
output.

The resource shrink bytecode visitor would visit all the field values
in these R classes and therefore find all the resources and treat them
as used. This meant a lot of resources that were otherwise eligible for
resource shrinking would not be removed.

This CL fixes that by deliberately skipping R$type classes from the
bytecode visitor.

It also adds more logging (for the --debug flag) such that developers
can track down which class is referencing a given resource id that
they believe should be unused.

Change-Id: I1fe71a17d4314735a983f707a6647c183b7bc7a3