81728: Support private resources in Gradle libraries
This CL allows resource files in Gradle projects to specify <public>
nodes in value resources, specifying the name and type of a resource
to designate as public. Here's an example:

<resources>
    <public name="mylib_app_name" type="string"/>
    <public name="mylib_public_string" type="string"/>
</resources>

As with all other value resources, you can define these in any file,
but the convention we’ll encourage via the new library template is to
place these in a file named "public.xml".

At build time, these nodes are used to create a public.txt file in the
AAR which declares which resources are considered public and visible
for clients of the library. This file has a very simple syntax:
type + space + name + newline. For the example above, it would be

string mylib_app_name
string mylib_public_string

If an AAR file contains a public.txt file, then all resources except
those specified in public.txt are considered to be private. Note that
the file can exist but be empty: that means that all resources are
considered private (and to declare this, create an empty public node
like this: <public />).

The public.txt file only contains public declarations for the current
library, not the transitive closure of the public symbols available
from libraries that this library depends on.

In the future, we can use the visibility information for resources to
(1) filter out private resources in places like code completion,
(2) warn from lint when these resources are accidentally being
    overridden or accidentally clashing between two separate libraries
(3) maybe even use the resource shrinking machinery to rename private
    resources within the library such that we can allow name clashes
    between private resources

Change-Id: I4c44c07c0386aef65dd6f618fbe286a4318f413b