Make DEBUG_INNER_TEST work when checking the model.
Since some of our integration tests (e.g. ModelTest) generate test projects on the fly (without distinct names), this CL also changes the behavior of DEBUG_INNER_TEST to kick in if it was set to any non-empty string. Filtering tests has to be done with --tests of -Dtest.single.

Change-Id: Ibab8b7843b3df904d24c14569624bb3fa4b1ec10