Fix resource resolution due to namespace conflict. [DO NOT MERGE]
If a resouce id name exists in the framework namepace, but not in the
project namespace (or it may be an id that hasn't been created yet),
sometimes the framework resource was returned instead of an error or
null.

This change also skips a warning which is logged when referencing a
style attribute that is not defined in the current configuration. For
example, if a widget defines its style as "style=?attr/foo" and foo is
not defined in the current theme. This is because even though it's an
error in most cases, the framework doesn't throw an error. And more
importantly, the support library (appcompat) does this. Showing a
warning in this case will confuse more users than it will help.

Change-Id: I89db6cc05014969ad38c8870544b68be352c2a7c
(cherry picked from commit 631827f251a67ef2a0c69b89d0a2b72c7feb8d87)