Removed all jniFolders when pure splits are enabled
While working on task parallelization, I realized the combined abi
and density pure split test was not running.
After adding it, I realized that although the splits were correct,
the main APK was still containing all the .so files.

Basically, when there are abi pure splits, we do not provide
any jniFolders to the main APK package task as all jni resources
are packaged in their respective split apk.

also added a missing reset() the build.gradle for the test.

Change-Id: Ie1d9d910db796ca943caeea5a307f1dd6acda005