Initial check-in of build profiling tool.
The critical path analysis will be included in a follow-up changelist.
We should also refine the model and decouple it from the view soon:
Chrome trace data was an easy prototyping choice, but doesn't allow queries,
etc.

I think for custom events, simply calling writeStart and writeEnd should do
the right thing (the listener could be a singleton).

Change-Id: I8db12832b179e4fbf87a6cb299fc5ad9b8e1ba48