Move flags from SessionParams to RenderParams
This change moves the map of property flags from SessionParams to
RenderParams. This is done so that DrawableParams can also use the map
to add arbitrary properties.

Unfortunately, this still means that the API version needs to be
updated. However, this makes it less likely to require API changes in
the future.

This also renames SessionParamsFlags to RenderParamsFlags.

Change-Id: I41aa375832a415a66681d7144ef6a139063cae73