Check all support libraries for obsolete versions.
Earlier, only appcompat-v7 and support-v4 were flagged for obsolete
versions. Other support libraries were not flagged in the editor when
the user was editing the files. This change removes the restriction from
just the two and assumes that all libraries with com.android.support as
the groupId are present in the maven repository that is shipped with the
SDK.

Change-Id: Ic4303e325a93541a803ae7b84b9fab8765d2fb0e