Allow lint to set error-range on name or value portion of attribute
Until now, if a detector flags an XML attribute with an error, the
whole attribute (prefix:name="value") is underlined. This CL adds a
couple of new location factory methods on XmlContext such that
detectors can associate an error with just the name portion of the
attribute, or just the value portion of the attribute. The CL also
updates a couple of detectors (and their unit tests) with this new
narrower scope.

It also fixes some issue explanations that were missing a newline
where the intention had been to create a new paragraph.

Change-Id: I2fa1bacb4eb863acb35be1c9085ec9c4752eca60