Delayed setting tested app mapping file.
Probably a regression when switching to proguard 5.2.1.
When the tested application does not request obfuscation, a mapping file
is not produced and setting it when invoking proguard on the test application
is now causing an exception.

Storing the mapping file location during task configuration and only
setting it before invoking the proguard task execution ensures the file
presence or absence.

Change-Id: I35526e0c88f3f1233f896d6bf90b306dbbc77ead