Improve API check: Check implicit casts, deprecation data, etc
This CL adds two user visible features:

(1) The API check now enforces both implicit and explicit casts, for
    both interfaces and super classes.  For example,
    android.view.KeyEvent only extends android.view.InputEvent since
    API 9. Therefore, while passing a KeyEvent to an API call (or
    assigning to a variable etc) which expects an InputEvent will
    compile just fine on recent APIs, this can lead to a runtime
    crash.

    The API database now records all the interface and superclass
    changes across API levels, and the ApiDetector now uses this to
    check implicit casts in variable declarations and in assignments,
    as well as in method call arguments.

(2) The API database now records all the deprecation versions for
    classes, methods and fields, such that it can be efficiently
    queried later. This will be used in Studio to suppress deprecation
    messages for APIs that were deprecated in a higher API level than
    the app's minSdkVersion.

To support this, the API binary database was changed quite a bit.
Even though it stores more information than before (interface and
superclass changes, as well as deprecation data), it should be a lot
smaller (about half the old size) as well as faster to search.

The speedup is achieved by breaking up the data into separate binary
search tables for packages, and once the package is known to only
binary search among classes in that package, and once the class is
known to members within that class. It's also smarter about pruning
members; instead of just pruning APIs with api level 1, it now prunes
all member entries where the API level is the same as the class API
level (unless the item needs to be kept to record deprecation data.)

There's also a comprehensive unit test which makes sure the database
can find and return the right result for every single class, method
and field; turns out that didn't pass in the old API lookup for some
corner cases which are now fixed.

(Also fixes a bug where type resolution wasn't working for ECJ local
variable entries.)

Fixes
https://code.google.com/p/android/issues/detail?id=174535
https://code.google.com/p/android/issues/detail?id=191120

Change-Id: Icf4a8d72ab161c3fd1024ec69d866698c5428dfd