Relax requirements about adb path.
In a previous CL, the adb path passed in via createBridge() call
was required to be a full path pointing to the executable. However,
some consumers of the API relied on adb being present on the path
and simply passing in "adb" as the parameter to the method.

This CL simply relaxes the unnecessary restriction that was
introduced.

Change-Id: Id38162ea9b9f67f7a6ad676562a9770aa49f287e