Move createParser to a new callback.
Move createParser() from LayoutlibCallback to a new class. Without this,
LayoutLib had to hold on to a the LayoutlibCallback object to get a
parser and this was a source of potential leaks. The change enables
LayoutLib to hold onto only the new callback, which should be a much
smaller object.

Change-Id: Ic1eae456827dad9937c4126849470e8756f7a162