Dominator computation now always excludes weak references.
Dominator computation now excludes weak/soft/phantom/finalized
references in its dominator computation.

Weak references are also now taken care of in the shortest distance
visitor for depth computation. When a link whose parent is a weak
reference and the child is a non-weak reference, the instance will not
be added to the frontier computation for depth purposes.

Added various refactoring for better naming.

Change-Id: I57016ce346738c63aaf0efdbed54071c11d00756