Support variant specific NDK options.
NDK options such as C flags specified in BuildTypes and ProductFlavors are now
merged and used to configure the native binaries.

ProductFlavorCombo is modified to support managed.ProductFlavor, which does
not inherit from model.ProductFlavor.  A Groupable interface is created to
support this.

getNdkConfig is rename to getNdk to simplify DSL.

core.NdkConfig is renamed to CoreNdkOptions for consistency.

Change-Id: Ifea600b6b7e6ab33fb1917f4f266ff6fb98be474