ddmlib DeviceMonitor: Cleanup comparison of device lists
Whenever the device state or list of devices changes, the device
monitor gets an update. On these updates, the monitor proceeds
to do different actions depending on whether devices were
added/removed or updated.

This CL separates out the logic to detect whether a device was
added/removed/updated, and the actions to be taken in those cases.

Change-Id: I45573437148b2c91ac63cfec1e2846195ecf244b