180795: @SuppressLint("HandlerLeak") don't suppress lint
We need to supply the AST node with the error such that the code which walks
the AST looking for suppress annotations can find it. (This detector used to
be a bytecode based detector so it didn't have an AST to walk; I should have
added this when porting it over to being AST based.) Caught a couple of other
similar report calls.

Change-Id: Id67c7740b6bbafaeab63519dee05e633a91bca4d