Fix the ls -l pattern to match POSIX.
toolbox's ls never included the nlink field, but POSIX mandates it, so toybox
ls has it.

Also add tests to check that we can support both old and new formats.

Also remove the LS_LD_PATTERN which contained a mistake and wasn't matching
anything. (There was either nothing to match the filename, or an accidental
final $, depending on what the intention was.)

Bug: http://b/23040458
Change-Id: Ibaa35763ab7f9bdf65915bdcc072760c46a297a2