Fix and expose stack trace info in perflib.
Previously stack traces were associated with the current heap at the
time the stack traces were loaded from the hprof file. This caused
stack traces not to be found if the current heap had changed since
loading the stack trace.

This commit stores all stack traces in the snapshot instead of
associating them with heaps. It also provides accessors for getting
the stack trace of an object, accessing the frames of a stack trace,
and accessing the components of a stack frame.

Change-Id: I0d27be07bb181ed87e0624e29e4e6b60e91d95bf