183500: Lint API check doesn't enforce Android M APIs
The root cause for this bug was that the platform-tools API
database did not include Android M APIs, and this was because
the API generator *only* looks at platform prebuilts.

This CL updates it such that we can pass in the currently
built android.jar, as well as the current version number,
and have that included in the database as well. That way,
we don't have to make sure a prebuilt is checked in first.

Change-Id: Ie6ce6881fac3c3df39c8976c854395432c4f0c4d