In the runtime, prune .dex files containing hidden classes
Over time, as you make repeated edits to the same file,
older .dex patches (which are all added to the class loader
at startup) end up providing classes that are no longer
reachable: all their classes are found in later patch files
that are added earlier in the class loader dex file list.

This CL records an index of the classes contained in each
patch, and at startup it will "compact" the patches together
such that only the patches containing unique, reachable
classes remain.

It also prunes the temporary dex files (the ones used for
BCI class loading; these have to exist on disk since that's
required by the DexClassLoader, but they do not need to
persist across IDE sessions.

Change-Id: I9d7e5ce19ac1ab48c534694bfd5d97f96478e1e1