Change Transform API to handle different types.
We started with a single Transform interface able
to handle all the types of transforms (noop, combined,
as_input). However this doesn't scale and fork_input
showed the first crack in the API. To allow us to
handle any new type, we're moving to a per-type
interface just for the transform() method.

Fork_input specific changes:
- consume the stream it processes.
- restrict to a single content type as input.

Also improved testing coverage:
- now test that the task is properly configured after
  the transform is added to the manager.
- now test that the transform receives the right data
  when the task converts from TStreams to TInput/Output.

Change-Id: If6d14fae8eeaca3e541524230dd92044bd596565